move html generate func to scripts will be used in the generation of htmls from the sub-category focus on SEO
