+++ repo_test.go
@@ -2,20 +2,13 @@ package main
 
 import (
 	"io/ioutil"
-	"os"
 	"regexp"
 	"sort"
 	"strings"
 	"testing"
-	"text/template"
 
 	"github.com/PuerkitoBio/goquery"
-	gfm "github.com/shurcooL/github_flavored_markdown"
 )
-
-type content struct {
-	Body string
-}
 
 func TestAlpha(t *testing.T) {
 	query := startQuery()
@@ -76,7 +69,7 @@ func TestSeparator(t *testing.T) {
 	}
 }
 func TestGenerateHTML(t *testing.T) {
-	err := generateHTML()
+	err := GenerateHTML()
 	if err != nil {
 		t.Errorf("html generate error '%s'", err.Error())
 	}
@@ -112,18 +105,4 @@ func checkAlphabeticOrder(t *testing.T, s *goquery.Selection) {
 	if t.Failed() {
 		t.Logf("expected order is:\n%s", strings.Join(sorted, "\n"))
 	}
-}
-
-func generateHTML() (err error) {
-	// options
-	readmePath := "./README.md"
-	tplPath := "tmpl/tmpl.html"
-	idxPath := "tmpl/index.html"
-	input, _ := ioutil.ReadFile(readmePath)
-	body := string(gfm.Markdown(input))
-	c := &content{Body: body}
-	t := template.Must(template.ParseFiles(tplPath))
-	f, err := os.Create(idxPath)
-	t.Execute(f, c)
-	return
 }
