+++ scripts.go
@@ -4,8 +4,12 @@ import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
+	"os"
+	"text/template"
 
 	"github.com/PuerkitoBio/goquery"
+	"github.com/gomarkdown/markdown"
+	"github.com/gomarkdown/markdown/parser"
 	"github.com/russross/blackfriday"
 )
 
@@ -26,4 +30,26 @@ func startQuery() *goquery.Document {
 		panic(err)
 	}
 	return query
+}
+
+type content struct {
+	Body string
+}
+
+func GenerateHTML() (err error) {
+	// options
+	readmePath := "./README.md"
+	tplPath := "tmpl/tmpl.html"
+	idxPath := "tmpl/index.html"
+	input, _ := ioutil.ReadFile(readmePath)
+	extensions := parser.CommonExtensions | parser.AutoHeadingIDs | parser.LaxHTMLBlocks
+	parser := parser.NewWithExtensions(extensions)
+
+	body := string(markdown.ToHTML(input, parser, nil))
+	// body := string(gfm.Markdown(input))
+	c := &content{Body: body}
+	t := template.Must(template.ParseFiles(tplPath))
+	f, err := os.Create(idxPath)
+	t.Execute(f, c)
+	return
 }
